Change length params to unsigned
Can't push or pop a -ve number of elements ;-)

+++ src/lib/FIFO/FIFO.cpp
@@ -57,7 +57,7 @@ void ICACHE_RAM_ATTR FIFO::push(const uint8_t data)
     }
 }
 
-void ICACHE_RAM_ATTR FIFO::pushBytes(const uint8_t *data, int len)
+void ICACHE_RAM_ATTR FIFO::pushBytes(const uint8_t *data, uint8_t len)
 {
     if (numElements + len > FIFO_SIZE)
     {
@@ -89,7 +89,7 @@ uint8_t ICACHE_RAM_ATTR FIFO::pop()
     }
 }
 
-void ICACHE_RAM_ATTR FIFO::popBytes(uint8_t *data, int len)
+void ICACHE_RAM_ATTR FIFO::popBytes(uint8_t *data, uint8_t len)
 {
     if (numElements < len)
     {
@@ -133,7 +133,7 @@ void ICACHE_RAM_ATTR FIFO::flush()
     numElements = 0;
 }
 
-bool ICACHE_RAM_ATTR FIFO::ensure(int requiredSize)
+bool ICACHE_RAM_ATTR FIFO::ensure(uint8_t requiredSize)
 {
     if(requiredSize > FIFO_SIZE)
         return false;
